requirements.txt
python-magic
python-magic-bin
ffmpeg-normalize